feat: add SpanAttributeUtils extension for SentrySpanV2
- Introduced a new extension `SpanAttributeUtils` to enhance `SentrySpanV2` by adding a method to conditionally add attributes if they are absent.
- Removed the unused import of `span_attribute_utils.dart` from `span_capture_pipeline.dart`.
- Created a new mock class `FakeSpanCapturePipeline` for testing purposes, which captures spans and scopes during tests.
