+++ packages/dart/lib/src/telemetry/span/span_capture_pipeline.dart
@@ -3,7 +3,6 @@ import 'package:meta/meta.dart';
 import '../../../sentry.dart';
 import '../../utils/internal_logger.dart';
 import '../default_attributes.dart';
-import 'span_attribute_utils.dart';
 import 'sentry_span_v2.dart';
 
 @internal
